Delete test_model (changing directory) & fix merge conflicts
