+++ tests/gordo_components/model/test_model.py
@@ -15,7 +15,7 @@ from sklearn.model_selection import cross_val_score, TimeSeriesSplit
 from keras.wrappers.scikit_learn import BaseWrapper
 
 from gordo_components.model import get_model
-from gordo_components.model.models import KerasLSTMAutoEncoder
+from gordo_components.model.models import KerasLSTMAutoEncoder, KerasLSTMForecast
 from gordo_components.model.factories import lstm_autoencoder
 from gordo_components.model.base import GordoBase
 from gordo_components.model.register import register_model_builder
@@ -64,7 +64,7 @@ def test_keras_autoencoder_crossval(model, kind):
     Model = pydoc.locate(f"gordo_components.model.models.{model}")
     model = Pipeline([("model", Model(kind=kind))])
 
-    X = np.random.random(size=8).reshape(-1, 2)
+    X = np.random.random(size=(15, 2))
 
     scores = cross_val_score(
         model, X, X, cv=TimeSeriesSplit(n_splits=2, max_train_size=2)
@@ -136,29 +136,65 @@ def test_save_load(model, kind):
 
 
 class KerasModelTestCase(unittest.TestCase):
-    def test__generate_window_valueerror(self):
-        # test for KerasAutoEncoder
-        # Assert that ValueError is raised if lookback_window > number of readings (rows of X)
+    def test_lookback_window_ae_valueerror_during_fit(self):
 
-        X = np.random.random(size=20).reshape(10, 2)
-        lookback_window = 11
+        # Assert that (for LSTMAutoEncoder) ValueError
+        # is raised in fit method if lookback_window > number of readings (rows of X)
+
+        lookback_window = 6
         with self.assertRaises(ValueError):
             model = KerasLSTMAutoEncoder(
                 kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
             )
-            gen = model._generate_window(X)
-            next(gen)
+            X = np.random.random(size=(10,))
+            model.fit(X)
 
-    def test__generate_window_output(self):
-        # test for lstm_autoencoder
-        # Check that right output is generated from _generate_window
+    def test_lookback_window_ae_valueerror_during_transform(self):
+
+        # Assert that (for LSTMAutoEncoder) ValueError
+        # is raised in fit method if lookback_window > number of readings (rows of X)
+
+        lookback_window = 3
+        model = KerasLSTMAutoEncoder(
+            kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
+        )
+        X_train = np.random.random(size=(4, 2))
+        model.fit(X_train)
+        with self.assertRaises(ValueError):
+            X_test = X_train[-3:-1, :]
+            model.transform(X_test)
+
+    def test_lookback_window_forecast_valueerror_during_fit(self):
+        # Assert that (for LSTMForecast) ValueError is raised
+        # in fit method if lookback_window >= number of readings (rows of X)
+        for lookback_window in [5, 6]:
+            model = KerasLSTMForecast(
+                kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
+            )
+            with self.assertRaises(ValueError):
+                X = np.random.random(size=(5, 2))
+                model.fit(X)
+
+    def test_lookback_window_forecast_valueerror_during_transform(self):
+        # Assert that (for LSTMForecast) ValueError is raised
+        # in fit method if lookback_window >= number of readings (rows of X)
+        X = np.random.random(size=(5, 2))
+        for lookback_window in [5, 6]:
+            model = KerasLSTMForecast(
+                kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
+            )
+            with self.assertRaises(ValueError):
+                model.fit(X)
+
+    def test_generate_window_lstm_ae_output(self):
+        # Check that right output is generated from generate_window (for KerasLSTMAutoEncoder)
         X = np.random.random(size=10).reshape(5, 2)
 
         lookback_window = 4
         model = KerasLSTMAutoEncoder(
             kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
         )
-        gen = model._generate_window(X)
+        gen = model.generate_window(X)
         gen_out_1 = next(gen)
         gen_out_2 = next(gen)
         self.assertEqual(gen_out_1[0].tolist(), X[0:4].reshape(1, 4, 2).tolist())
@@ -171,7 +207,7 @@ class KerasModelTestCase(unittest.TestCase):
         model = KerasLSTMAutoEncoder(
             kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
         )
-        gen_no_y = model._generate_window(X, output_y=False)
+        gen_no_y = model.generate_window(X, output_y=False)
         gen_no_y_out_1 = next(gen_no_y)
         gen_no_y_out_2 = next(gen_no_y)
         gen_no_y_out_3 = next(gen_no_y)
@@ -179,31 +215,61 @@ class KerasModelTestCase(unittest.TestCase):
         self.assertEqual(gen_no_y_out_2.tolist(), X[1:3].reshape(1, 2, 2).tolist())
         self.assertEqual(gen_no_y_out_3.tolist(), X[2:4].reshape(1, 2, 2).tolist())
 
-    def test_transform_output_dim(self):
-        # test for KerasLSTMAutoEncoder
-        # test dimension of output
+    def test_generate_window_lstm_forecast_output(self):
+        # Check that right output is generated from generate_window (for KerasLSTMForecast)
+        X = np.random.random(size=(5, 2))
 
-        X_train = np.random.random(size=15).reshape(5, 3)
+        lookback_window = 3
+        model = KerasLSTMForecast(
+            kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
+        )
+        gen = model.generate_window(X)
+        gen_out_1 = next(gen)
+        gen_out_2 = next(gen)
+        self.assertEqual(gen_out_1[0].tolist(), X[0:3].reshape(1, 3, 2).tolist())
+        self.assertEqual(gen_out_2[0].tolist(), X[1:4].reshape(1, 3, 2).tolist())
+        self.assertEqual(gen_out_1[1].tolist(), X[3].reshape(1, 2).tolist())
+        self.assertEqual(gen_out_2[1].tolist(), X[4].reshape(1, 2).tolist())
+
+        X = np.random.random(size=(4, 2))
+        lookback_window = 2
+        model = KerasLSTMForecast(
+            kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
+        )
+        gen_no_y = model.generate_window(X, output_y=False)
+        gen_no_y_out_1 = next(gen_no_y)
+        gen_no_y_out_2 = next(gen_no_y)
+        self.assertEqual(gen_no_y_out_1.tolist(), X[0:2].reshape(1, 2, 2).tolist())
+        self.assertEqual(gen_no_y_out_2.tolist(), X[1:3].reshape(1, 2, 2).tolist())
+
+    def test_lstmae_transform_output(self):
+        # test for KerasLSTMAutoEncoder
+        #  - test dimension of output
+        #  - test that first half of output is testing data
+
+        X_train = np.random.random(size=(5, 3))
         lookback_window = 3
         model = KerasLSTMAutoEncoder(
             kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
         )
         model = model.fit(X_train)
-        X_test = np.random.random(size=12).reshape(4, 3)
+        X_test = np.random.random(size=(4, 3))
         out = model.transform(X_test)
         self.assertEqual(out.shape, (2, 6))
+        self.assertEqual(out[:, :3].tolist(), X_test[lookback_window - 1 :, :].tolist())
 
-    def test_transform_output(self):
-        # test for KerasLSTMAutoEncoder
-        # test that first half of output is testing data
+    def test_lstmforecast_transform_output(self):
+        # test for KerasLSTMForecast
+        #  - test dimension of output
+        #  - test that first half of output is testing data
 
-        X_train = np.random.random(size=15).reshape(5, 3)
+        X_train = np.random.random(size=(5, 3))
         lookback_window = 3
-        model = KerasLSTMAutoEncoder(
+        model = KerasLSTMForecast(
             kind=lstm_autoencoder.lstm_model, lookback_window=lookback_window
         )
         model = model.fit(X_train)
-        X_test = np.random.random(size=12).reshape(4, 3)
+        X_test = np.random.random(size=(6, 3))
         out = model.transform(X_test)
-        self.assertEqual(out[0, :3].tolist(), X_test[2, :].tolist())
-        self.assertEqual(out[1, :3].tolist(), X_test[3, :].tolist())
+        self.assertEqual(out.shape, (3, 6))
+        self.assertEqual(out[:, :3].tolist(), X_test[lookback_window:, :].tolist())
